requirements.txt
fastapi
pydantic
python-dotenv
pandas
uvicorn
cProfile
pstats
